+++ crates/dispatch2/CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Changed how memory management works to match other `objc2` crates. Instead
   of types like `DispatchGroup` holding the retain count internally, it is now
   done externally by `DispatchRetained<DispatchGroup>`.
+- Converted `DispatchObject` to a trait.
 - Renamed dispatch types to be prefixed with `Dispatch` (e.g. `DispatchGroup`,
   `DispatchQueue`, `DispatchOnce` etc).